requirements.txt
python-telegram-bot==21.11.1
Flask==3.1.0
psycopg[binary]==3.2.9
